src/requirements.txt
anyio==3.6.2
attrs==22.1.0
beanie==1.15.4
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
dnspython==2.2.1
exceptiongroup==1.0.4
fastapi==0.87.0
flake8==6.0.0
h11==0.14.0
httpcore==0.16.2
httpx==0.23.1
idna==3.4
iniconfig==1.1.1
isort==5.10.1
mccabe==0.7.0
mongomock==4.1.2
mongomock-motor==0.0.13
motor==3.1.1
multidict==6.0.2
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.2
platformdirs==2.5.4
pluggy==1.0.0
pycodestyle==2.10.0
pydantic==1.10.2
pyflakes==3.0.1
pymongo==4.3.3
pyparsing==3.0.9
pytest==7.2.0
pytest-asyncio==0.20.2
python-dotenv==0.21.0
requests==2.28.1
rfc3986==1.5.0
sentinels==1.0.0
sniffio==1.3.0
starlette==0.21.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
yarl==1.8.1
